Aloysius

+++ DIPAvatarChat/Assets/Script/Managers/APIs/ConversationBackendManager.cs
@@ -181,100 +181,7 @@ public class ConversationBackendManager : Singleton<ConversationBackendManager>
         }
     }
 
-    public bool AddConversationMember(string conversationID, string memberEmail)
-    {
-
-        db = FirebaseFirestore.DefaultInstance;
-        _userPath = AuthManager.Instance.userPathData;
-
-        if (db == null)
-        {
-            Debug.LogError("Firebase Firestore is not initialized. Make sure it's properly configured.");
-            return false;
-        }
-
-        try
-        {
-            // Get a reference to the conversation document using the provided conversation ID
-            DocumentReference conversationRef = db.Collection("conversation").Document(conversationID);
-
-            // Use FieldValue.ArrayUnion to add a member to the "members" array field
-            Dictionary<string, object> updateData = new Dictionary<string, object>
-        {
-            { "members", FieldValue.ArrayUnion(memberEmail) }
-        };
-
-            // Update the conversation document to add the member
-            conversationRef.UpdateAsync(updateData)
-                .ContinueWithOnMainThread(task =>
-                {
-                    if (task.IsCompleted)
-                    {
-                        Debug.Log("Member added to the conversation successfully.");
-                    }
-                    else if (task.IsFaulted)
-                    {
-                        Debug.LogError("Error adding member to the conversation: " + task.Exception);
-                    }
-                });
-
-            return true;
-        }
-        catch (Exception e)
-        {
-            Debug.LogError("Error adding member to the conversation: " + e.Message);
-            return false;
-        }
-    }
-
-
-    public bool DeleteConversationMember(string conversationID, string memberEmail)
-    {
-        db = FirebaseFirestore.DefaultInstance;
-        _userPath = AuthManager.Instance.userPathData;
-
-        if (db == null)
-        {
-            Debug.LogError("Firebase Firestore is not initialized. Make sure it's properly configured.");
-            return false;
-        }
-
-        try
-        {
-            // Get a reference to the conversation document using the provided conversation ID
-            DocumentReference conversationRef = db.Collection("conversation").Document(conversationID);
-
-            // Use FieldValue.ArrayRemove to remove the member from the "members" array field
-            Dictionary<string, object> updateData = new Dictionary<string, object>
-        {
-            { "members", FieldValue.ArrayRemove(memberEmail) }
-        };
-
-            // Update the conversation document to remove the member
-            conversationRef.UpdateAsync(updateData)
-                .ContinueWithOnMainThread(task =>
-                {
-                    if (task.IsCompleted)
-                    {
-                        Debug.Log("Member removed from the conversation successfully.");
-                    }
-                    else if (task.IsFaulted)
-                    {
-                        Debug.LogError("Error removing member from the conversation: " + task.Exception);
-                    }
-                });
-
-            return true;
-        }
-        catch (Exception e)
-        {
-            Debug.LogError("Error removing member from the conversation: " + e.Message);
-            return false;
-        }
-    }
-
-
-    public bool DeleteConversation(string conversationID)
+    public async Task<bool> DeleteConversationTask(string conversationID)
     {
         try
         {
@@ -283,19 +190,26 @@ public class ConversationBackendManager : Singleton<ConversationBackendManager>
             // Get a reference to the conversation document using the provided conversation ID
             DocumentReference conversationRef = db.Collection("conversation").Document(conversationID);
 
-            // Delete the conversation document
-            conversationRef.DeleteAsync().ContinueWithOnMainThread(task =>
-            {
-                if (task.IsCompleted)
-                {
-                    Debug.Log("Conversation deleted successfully.");
-                }
-                else if (task.IsFaulted)
-                {
-                    Debug.LogError("Error deleting conversation: " + task.Exception.ToString());
-                }
-            });
+            DocumentSnapshot convSnapshot = await GetConversationByIDTask(conversationID);
+            ConversationData convData = convSnapshot.ConvertTo<ConversationData>();
+            List<string> convMembers = new List<string>(convData.members);
 
+            if (await DeleteAllMessageTask(convData.conversationID))
+            {
+                // Delete the conversation document
+                conversationRef.DeleteAsync().ContinueWithOnMainThread(task =>
+                {
+                    if (task.IsCompleted)
+                    {
+                        Debug.Log("Conversation deleted successfully.");
+                        DeleteConversationFromUsers(convMembers[0], convMembers[1], convData.conversationID);
+                    }
+                    else if (task.IsFaulted)
+                    {
+                        Debug.LogError("Error deleting conversation: " + task.Exception.ToString());
+                    }
+                });
+            }
             return true; // Return true to indicate that the deletion process has started.
         }
         catch (Exception e)
@@ -305,6 +219,80 @@ public class ConversationBackendManager : Singleton<ConversationBackendManager>
         }
     }
 
+    public async Task<bool> DeleteAllMessageTask(string convID)
+    {
+        CollectionReference collectionReference = db.Collection("message");
+
+        try
+        {
+            collectionReference.GetSnapshotAsync().ContinueWith(task =>
+            {
+                if (task.IsFaulted)
+                {
+                    Debug.LogError("Error getting documents: " + task.Exception);
+                }
+
+                foreach (DocumentSnapshot documentSnapshot in task.Result.Documents)
+                {
+                    if (documentSnapshot.Exists)
+                    {
+                        DocumentReference msgRef = db.Collection("message").Document(documentSnapshot.Id);
+                        MessageData messageData = documentSnapshot.ConvertTo<MessageData>();
+
+                        if (messageData.conversationID == convID)
+                        {
+                            // Delete the conversation document
+                            msgRef.DeleteAsync().ContinueWithOnMainThread(task =>
+                            {
+                                if (task.IsCompleted)
+                                {
+                                    Debug.Log("Message deleted successfully.");
+                                }
+                                else if (task.IsFaulted)
+                                {
+                                    Debug.LogError("Error deleting msg: " + task.Exception.ToString());
+                                }
+                            });
+                        }
+                    }
+                }
+            });
+            return true;
+        }
+        catch (Exception e)
+        {
+            Debug.LogError("Error deleting conversation: " + e.Message);
+            return false;
+        }
+    }
+
+    public async void DeleteConversationFromUsers(string myEmail, string theirEmail, string conversation)
+    {
+        DocumentSnapshot myUserDoc = await UserBackendManager.Instance.GetUserByEmailTask(myEmail);
+        UserData myUserData = myUserDoc.ConvertTo<UserData>();
+
+        DocumentSnapshot theirUserDoc = await UserBackendManager.Instance.GetUserByEmailTask(theirEmail);
+        UserData theirUserData = theirUserDoc.ConvertTo<UserData>();
+
+        List<string> myConversationsList = new List<string>(myUserData.conversations);
+        List<string> theirConversationsList = new List<string>(theirUserData.conversations);
+
+        myConversationsList.Remove(conversation);
+        theirConversationsList.Remove(conversation);
+
+        Dictionary<string, object> myUserDict = new Dictionary<string, object>
+        {
+            { "conversations", myConversationsList }
+        };
+
+        Dictionary<string, object> theirUserDict = new Dictionary<string, object>
+        {
+            { "conversations", theirConversationsList }
+        };
+
+        db.Document("user/" + myEmail).UpdateAsync(myUserDict);
+        db.Document("user/" + theirEmail).UpdateAsync(theirUserDict);
+    }
 
     public ConversationData DictionaryToConversationData(Dictionary<string, object> firestoreData, List<string> members, List<string> messages)
     {
